% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implausible_values.R
\name{check_implausible}
\alias{check_implausible}
\title{Check plausibility of doses and medication quantities}
\usage{
check_implausible(data = NULL, min_max_dat = NULL)
}
\arguments{
\item{data}{a data frame object containing the prescription information}

\item{min_max_dat}{a data frame object containing minimum and maximum possible values for each drug}
}
\value{
A data frame with the same structure as the input, plus two columns:
\itemize{
\item{\code{implausible_qty}}
\item{\code{implausible_ndd}}
},
both logical vectors indicating whether, respectively, the drug quantity or
numeric daily dose are outside the plausible ranges given by \code{min_max_dat}.
If there is no \code{prodcode} match to \code{min_max_dat}, plausibility is
given as \code{NA}.
}
\description{
This function flags quantity and ndd data that are outside specified acceptable ranges.
The acceptable level may be obtained from NICE BNF, or provided manually.
}
