% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ui.R
\name{clean_duration}
\alias{clean_duration}
\title{Clean implausibly-long prescription durations}
\usage{
clean_duration(data, max_months = Inf, method = c("truncate", "remove"))
}
\arguments{
\item{data}{A data frame containing a column called \code{duration}}

\item{max_months}{The maximum plausible prescription length in months}

\item{method}{Either 'truncate' or 'remove'.}
}
\description{
Given a prescription length limit, truncate any prescriptions that appear to
be longer than this, or mark them as missing.
}
\details{
If you want to 'do nothing', simply set \code{max_months} to an arbitrarily
high number. By default, the maximum is infinite, so nothing should happen.
}
\examples{
long_presc <- data.frame(duration = c(100, 300, 400, 800))
clean_duration(long_presc, 6)
clean_duration(long_presc, 12, 'remove')

}
