% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec10_gap_bn_prescription.R
\name{dec10_gap_bn_prescription}
\alias{dec10_gap_bn_prescription}
\title{dec10: Handle sequential prescriptions with short gaps
options are	10a	Do nothing
change stop date of first prescription to start date of next if gap is <=
10b_15  15 days
10b_30  30 days
10b_60  60 days}
\usage{
dec10_gap_bn_prescription(dataset1, decision)
}
\arguments{
\item{dataset, }{decsion}
}
\value{
dataset
}
\description{
dec10: Handle sequential prescriptions with short gaps
options are	10a	Do nothing
change stop date of first prescription to start date of next if gap is <=
10b_15  15 days
10b_30  30 days
10b_60  60 days
}
