% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec2_missing_qty}
\alias{dec2_missing_qty}
\title{Missing quantity}
\usage{
dec2_missing_qty(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containing prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"2a"}{Leave as missing (implicitly drop this prescription)}
\item{"2b1"}{set to mean for individual's prescriptions for that drug}
\item{"2b2"}{set to mean for practice's prescriptions for that drug}
\item{"2b3"}{set to mean for populations's prescriptions for that drug}
\item{"2c1"}{set to median for individual's prescriptions for that drug}
\item{"2c2"}{set to median for practice's prescriptions for that drug}
\item{"2c3"}{set to median for population's prescriptions for that drug}
\item{"2d1"}{set to mode for individual's prescriptions for that drug}
\item{"2d2"}{set to mode for practice's prescriptions for that drug}
\item{"2d3"}{set to mode for population's prescriptions for that drug}
\item{"2e1"}{use value of individual's next prescription}
\item{"2e2"}{use value of practice's next prescription}
\item{"2e3"}{use value of population's next prescription}
\item{"2f1"}{use value of individual's previous prescription}
\item{"2f2"}{use value of practice}'s previous prescription
\item{"2f3"}{use value of population's previous prescription}
}}
}
\value{
Dataframe with the same structure as the input
}
\description{
This function accepts a data frame containing prescription information and
replace the missing quntity by one of the following options. The input data
must have the following columns: patid, prodcode, pracid (if practice level
information is used for replacement) and qty.
}
