% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec6_select_stop_date}
\alias{dec6_select_stop_date}
\title{Select stop date}
\usage{
dec6_select_stop_date(data = NULL, decision)
}
\arguments{
\item{data}{a data frame containing prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"6a"}{stop1 (start + numdays)}
\item{"6b"}{stop2 (start + dose_duration)}
\item{"6c"}{stop3 (start + qty/ndd)}
}}
}
\value{
Dataframe with the same structure as the input
}
\description{
This function accepts a data frame containing prescription information already processed by
dec5_clean_duration, and select which stop date to use from one of the following options.
This step can only be called after dec5_clean_duration is executed.  The input dataset
must have the following columns: event_date, numdays, dose_duration, new_duration (column
name created by dec5_clean_duration).
}
