% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec9_overlaping_prescription.R
\name{dec9_overlaping_prescription}
\alias{dec9_overlaping_prescription}
\title{Handle overlapping prescriptions
options are:	9a  do nothing, allow prescriptions to overlap (implicity sum doses)
9b  move later prescription to next available time that this product is not prescribed}
\usage{
dec9_overlaping_prescription(dataset1, decision)
}
\arguments{
\item{dataset, }{decsion}
}
\value{
dataset
}
\description{
Handle overlapping prescriptions
options are:	9a  do nothing, allow prescriptions to overlap (implicity sum doses)
9b  move later prescription to next available time that this product is not prescribed
}
