% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{decision_3}
\alias{decision_3}
\title{Decision 3: impute implausible daily doses}
\usage{
decision_3(data, decision = "a")
}
\arguments{
\item{data}{a data frame}

\item{decision}{one of the following strings:
\describe{
\item{"a"}{do nothing; leave implausible values as-is}
\item{"b"}{set implausible values to missing}
\item{"c1"}{set to mean for individual's prescriptions for that drug}
\item{"c2"}{set to mean for practice's prescriptions for that drug}
\item{"c3"}{set to mean for populations's prescriptions for that drug}
\item{"d1"}{set to median for individual's prescriptions for that drug}
\item{"d2"}{set to median for practice's prescriptions for that drug}
\item{"d3"}{set to median for population's prescriptions for that drug}
\item{"e1"}{set to mode for individual's prescriptions for that drug}
\item{"e2"}{set to mode for practice's prescriptions for that drug}
\item{"e3"}{set to mode for population's prescriptions for that drug}
\item{"f1"}{use value of individual's next prescription}
\item{"f2"}{use value of practice's next prescription}
\item{"f3"}{use value of population's next prescription}
\item{"g1"}{use value of individual's previous prescription}
\item{"g2"}{use value of practice's previous prescription}
\item{"g3"}{use value of population's previous prescription}
}}
}
\description{
A light wrapper around \code{\link{impute_ndd}}.
}
\note{
Decisions \code{f} and \code{g} are not yet implemented.
}
\seealso{
Other decision functions: 
\code{\link{decision_10}()},
\code{\link{decision_1}()},
\code{\link{decision_2}()},
\code{\link{decision_4}()},
\code{\link{decision_5}()},
\code{\link{decision_6}()},
\code{\link{decision_7}()},
\code{\link{decision_8}()},
\code{\link{decision_9}()},
\code{\link{drug_prep}()}
}
\concept{decision functions}
