% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{decision_7}
\alias{decision_7}
\title{Decision 7: impute missing prescription durations}
\usage{
decision_7(data, decision = "a")
}
\arguments{
\item{data}{a data frame}

\item{decision}{one  of the following strings:
\describe{
\item{"a"}{Leave missing durations as-is (implicitly drop the prescription)}
\item{"b"}{Use mean prescription duration for that drug, for that individual}
\item{"c"}{Use mean prescription duration for that drug, for the population}
\item{"d"}{Use individual mean duration; if not available use population mean}
}}
}
\description{
A light wrapper around \code{\link{impute_duration}}.
}
\seealso{
Other decision functions: 
\code{\link{decision_10}()},
\code{\link{decision_1}()},
\code{\link{decision_2}()},
\code{\link{decision_3}()},
\code{\link{decision_4}()},
\code{\link{decision_5}()},
\code{\link{decision_6}()},
\code{\link{decision_8}()},
\code{\link{decision_9}()},
\code{\link{drug_prep}()}
}
\concept{decision functions}
