% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{decision_8}
\alias{decision_8}
\title{Decision 8: disambiguate prescriptions with the same start date}
\usage{
decision_8(data, decision = "a")
}
\arguments{
\item{data}{a data frame}

\item{decision}{one of the following strings
\describe{
\item{"a"}{do nothing}
\item{"b"}{replace with a prescription of duration equal to the mean}
\item{"c"}{choose the shortest prescription}
\item{"d"}{choose longest prescription}
\item{"e"}{replace with a prescription of duration equal to the sum}
}}
}
\description{
A light wrapper around \code{\link{impute_duration}}, followed by removing
duplicate rows with the same combination of \code{prodcode}, \code{patid}
and \code{start_date}.
}
\seealso{
Other decision functions: 
\code{\link{decision_10}()},
\code{\link{decision_1}()},
\code{\link{decision_2}()},
\code{\link{decision_3}()},
\code{\link{decision_4}()},
\code{\link{decision_5}()},
\code{\link{decision_6}()},
\code{\link{decision_7}()},
\code{\link{decision_9}()},
\code{\link{drug_prep}()}
}
\concept{decision functions}
