% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_therapy}
\alias{example_therapy}
\title{Example electronic prescription dataset}
\format{
An object of class \code{data.frame} with 30 rows and 6 columns.
}
\usage{
example_therapy
}
\description{
Based on a hypothetical 'therapy' file from the Clinical Practical Research
Datalink (CPRD), a UK database of primary care records.
}
\note{
This dataset is now generated deterministically, so it will not vary
between sessions.
}
\keyword{datasets}
