% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_Multiverse.R
\name{execute_Multiverse}
\alias{execute_Multiverse}
\title{Execute an analysis of a Multiverse data
These are functions which allow the user to execute parts or whole of the multiverse.}
\usage{
execute_Multiverse(formula, data, ores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{formula}{will be the same model formula to be implimenetd across multiple data}

\item{cores}{Indicates the number of cores to use. This will execute the entire multiverse in parallel.
Defaults to NULL (running in a single core)}
}
\value{
model
}
\description{
Execute an analysis of a Multiverse data
These are functions which allow the user to execute parts or whole of the multiverse.
}
