% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_universe.R
\name{execute_universe}
\alias{execute_universe}
\title{Execute an analysis of a single data generated}
\usage{
execute_universe(formula, data, ...)
}
\value{
model
}
\description{
Execute an analysis of a single data generated
}
