% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_Multiverse.R
\name{extract_Multiverse}
\alias{extract_Multiverse}
\title{Extract variables and objects from the Multiverse analysis
details to be add}
\usage{
extract_Multiverse()
}
\description{
Extract variables and objects from the Multiverse analysis
details to be add
}
