% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_universe.R
\name{extract_universe}
\alias{extract_universe}
\title{Extract variables and objects from the universe analysis
details to be add}
\usage{
extract_universe()
}
\description{
Extract variables and objects from the universe analysis
details to be add
}
