% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffreduce.R
\name{ffreduce}
\alias{ffreduce}
\title{freduce
modified version of reduce function in r}
\usage{
ffreduce(value, function_list, decision)
}
\arguments{
\item{value, }{function_list, decision}
}
\value{
value
}
\description{
freduce
modified version of reduce function in r
}
