% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_mode}
\alias{get_mode}
\title{Get the mode (most common value) of a vector}
\usage{
get_mode(v, na.rm = TRUE)
}
\arguments{
\item{v}{a vector}

\item{na.rm}{Logical. If \code{TRUE} (the default), find mode of non-\code{NA} values}
}
\description{
Get the mode (most common value) of a vector
}
