% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implausible_values.R
\name{Implausible_values}
\alias{Implausible_values}
\title{Define Implausible}
\usage{
Implausible_values(dataset1 = NULL, min_max_dat = NULL)
}
\arguments{
\item{dataset1}{a data frame object containing the prescription information}

\item{min_max_dat}{a data frame object containing minimum and maximum possible values for each drug}
}
\value{
Dataframe with the same structure as the input.
}
\description{
This function flag quantity and ndd informations that are beyond the acceptable levels in the dataset.
The acceptable level is obtained from NICE BNF.
}
