% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ui.R
\name{impute}
\alias{impute}
\title{Impute missing or implausible values}
\usage{
impute(
  data,
  variable,
  which_fun = is.na,
  fun = c("ignore", "mean", "median", "mode", "missing"),
  group = c("population", "patid", "pracid"),
  ...
)
}
\arguments{
\item{data}{A data frame containing columns \code{prodcode}, \code{pracid}, \code{patid}}

\item{variable}{Unquoted name of the column in \code{dataset} to be imputed}

\item{which_fun}{Function applied to \code{variable} that returns a logical vector indicating which elements to impute. Defaults to \code{\link[base:NA]{is.na}}}

\item{fun}{Method for imputing the values. See details.}

\item{group}{Level of structure for imputation. Defaults to whole study population.}

\item{...}{Not used}
}
\value{
A data frame of the same structure as \code{data}, with values imputed
}
\description{
Impute missing or implausible values
}
\details{
Possible values for \code{fun} are
\itemize{
\item \code{ignore}. Do nothing, leaving input unchanged.
\item \code{mean}. Replace values with the mean by \code{group}
\item \code{median}. Replace values with the median by \code{group}
\item \code{mode}. Replace values with the most common value by \code{group}
\item \code{missing}. Replace values with \code{NA}, to mark as missing.
}
}
\examples{
iris2 <- iris
iris2[sample(nrow(iris), 50), 'Petal.Length'] <- NA
colnames(iris2)[colnames(iris2) == 'Species'] <- 'prodcode'
iris2$pracid <- as.numeric(as.numeric(iris2$prodcode) < 2)
impute(iris2, Petal.Length, fun = 'median')
impute(iris2, Petal.Length, fun = 'mean', group = 'population')
impute(iris2, Petal.Length, fun = 'mean', group = 'pracid')

}
