% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ui.R
\name{impute_duration}
\alias{impute_duration}
\title{Replace missing durations}
\usage{
impute_duration(
  data,
  method = c("ignore", "mean", "median", "mode", "replace"),
  group = c("population", "patid", "pracid"),
  ...
)
}
\arguments{
\item{data}{A data frame containing columns \code{prodcode}, \code{pracid}, \code{patid}}

\item{method}{Method for imputing the values. See details.}

\item{group}{Level of structure for imputation. Defaults to whole study population.}

\item{...}{Extra arguments, currently ignored}
}
\description{
Instead of replacing missing stop dates, we impute the durations and then
infer the stop dates from there.
}
\examples{
example_duration <- transform(example_therapy, duration = qty / ndd)
impute_duration(example_duration, method = 'mean', group = 'patid')

}
