% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ui.R
\name{impute_ndd}
\alias{impute_ndd}
\title{Replace implausible or missing numerical daily doses (NDD)}
\usage{
impute_ndd(
  data,
  which_fun,
  fun = c("ignore", "mean", "median", "mode", "missing"),
  group = c("population", "patid", "pracid"),
  ...
)
}
\description{
Replace implausible or missing numerical daily doses (NDD)
}
