% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_ui.R
\name{impute_qty}
\alias{impute_qty}
\title{Find implausible entries
Replace implausible or missing prescription quantities}
\usage{
impute_qty(
  data,
  which_fun,
  fun = c("ignore", "mean", "median", "mode", "missing"),
  group = c("population", "patid", "pracid"),
  ...
)
}
\arguments{
\item{data}{A data frame containing columns \code{prodcode}, \code{pracid}, \code{patid}}

\item{which_fun}{Function applied to \code{variable} that returns a logical vector indicating which elements to impute. Defaults to \code{\link[base:NA]{is.na}}}

\item{fun}{Method for imputing the values. See details.}

\item{group}{Level of structure for imputation. Defaults to whole study population.}

\item{...}{Not used}
}
\description{
Find implausible entries
Replace implausible or missing prescription quantities
}
