% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{outside_range}
\alias{outside_range}
\title{Do values fall outside a specified 'plausible' range?}
\usage{
outside_range(x, lower, upper, open = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{lower}{minimum plausible value}

\item{upper}{maximum plausible value}

\item{open}{logical. If \code{TRUE}, values exactly equal to \code{lower} or \code{upper} are also considered implausible}
}
\description{
A utility function for indicating if elements of a vector are implausible.
}
\details{
Though the function \code{\link[dplyr]{between}} already exists, it is not vectorised over the bounds.
}
