% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Multiverse.R
\name{plot_Multiverse}
\alias{plot_Multiverse}
\title{This function will visualize the results of a Multiverse object returned by extract_Multiverse
details to be add}
\usage{
plot_Multiverse()
}
\description{
This function will visualize the results of a Multiverse object returned by extract_Multiverse
details to be add
}
