% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.MVD.R
\name{run.drugPREP.MVD}
\alias{run.drugPREP.MVD}
\title{run drug prep multiverse
This function allows the user to define multiple analysis options for a particular step
After creating permutation of each analysis option and step, it runs run.drugPREP
Since there will be thousends of analysis pathways
there is a need to distribute the computation
The user I think need to disignate a specific directory for storing the generated multiverse data}
\usage{
run.drugPREP.MVD(dataset1 = NULL, decisions = NULL)
}
\arguments{
\item{dataset, }{decisons,}
}
\value{
dataset
}
\description{
Not sure yet what is the best approch for this step
}
