% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_dosemine.R
\name{call_dosemine}
\alias{call_dosemine}
\title{Compute ndd
This function calls the dosemining r package and compute the ndd according to the suplemented decision}
\usage{
call_dosemine(dataset1 = NULL, decision = NULL)
}
\arguments{
\item{dataset1}{a data frame object containing the prescription information}

\item{decision}{a character vector for computing number of thing to take per day.}
}
\value{
dataset
}
\description{
Compute ndd
This function calls the dosemining r package and compute the ndd according to the suplemented decision
}
