% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ndd.R
\name{compute_ndd}
\alias{compute_ndd}
\title{Compute the number of daily dose (ndd)}
\usage{
compute_ndd(dataset1 = NULL, decision = NULL)
}
\arguments{
\item{dataset1}{a data frame object containing the prescription information.
The dataset is must have column names 'dosageid' and 'text'. This dataset is
expected to be a result of merging CPRD 'therapy' table with 'commondosage' table.}

\item{decision}{a character variable for computing number of (ndd).
\itemize{
\item{"min_min"}{ minimum dose number and minimum dose frequency}
\item{"min_mean"}{ minimum dose number and mean dose frequency}
\item{"min_max"}{ minimum dose number and maximum dose frequency}
\item{"mean_min"}{ mean dose number and minimum dose frequency}
\item{"mean_mean"}{ mean dose number and mean dose frequency}
\item{"mean_max"}{ mean dose number and maximum dose frequency}
\item{"max_min"}{ maximum dose number and minimum dose frequency}
\item{"max_mean"}{ maximum dose number and mean dose frequency}
\item{"max_max"}{ maximum dose number and maximum dose frequency}
}}
}
\value{
Dataframe with the same structure as the input with additional column named ndd
}
\description{
The function calls the doseminer r-package to extract doses from text prescription
and compute the (ndd) according to the given decision.
The general formula for computing (ndd) is given by (ndd=DF*DN/DI), where (DF) is dose
frequency, (DN) is dose number, and (DI) is dose interval. Prescriptions
can have a varible dose frequency or dose number such as '2-4 tablets'. In this case, the user can choose
minimum or maximum or average of value of DN or DF.
}
