% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec10_gap_bn_prescription}
\alias{dec10_gap_bn_prescription}
\title{Handle sequential prescriptions with short gaps}
\usage{
dec10_gap_bn_prescription(data = NULL, decision)
}
\arguments{
\item{data}{a data frame containing prescription information}

\item{decision}{a character specifying the decision to consider for processing
\itemize{
\item{"10a"}{do nothing}
\item{"10b"}{change stop date of first prescription to start date of next if gap is <= 15 days}
\item{"10c"}{change stop date of first prescription to start date of next if gap is <= 30 days}
\item{"10d"}{change stop date of first prescription to start date of next if gap is <= 60 days}
}}
}
\value{
A data frame with the same structure as the input
}
\description{
This function provides option for filling gaps between prescription of the same type
for an individual depending on the choosen decison.
}
