% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec10_gap_bn_prescription}
\alias{dec10_gap_bn_prescription}
\title{Handle sequential prescriptions with short gaps}
\usage{
dec10_gap_bn_prescription(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containg prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"10a"}{do nothing}
\item{"10b_15"}{change stop date of first prescription to start date of next if gap is <= 15 days}
\item{"10b_30"}{change stop date of first prescription to start date of next if gap is <= 30 days}
\item{"10b_60"}{change stop date of first prescription to start date of next if gap is <= 60 days}
}}
}
\value{
Dataframe with the same structure as the input
}
\description{
Handle sequential prescriptions with short gaps
}
