% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec1_implausible_qty}
\alias{dec1_implausible_qty}
\title{Implausible  quantity}
\usage{
dec1_implausible_qty(data = NULL, decision)
}
\arguments{
\item{data}{a data frame containing prescription information.}

\item{decision}{a character specifying the decison to consider for processing.
\itemize{
\item{"1a"}{use implausible value}
\item{"1b"}{set to missing}
\item{"1c1"}{set to mean for individual's prescriptions for that drug}
\item{"1c2"}{set to mean for practice's prescriptions for that drug}
\item{"1c3"}{set to mean for populations's prescriptions for that drug}
\item{"1d1"}{set to median for individual's prescriptions for that drug}
\item{"1d2"}{set to median for practice's prescriptions for that drug}
\item{"1d3"}{set to median for population's prescriptions for that drug}
\item{"1e1"}{set to mode for individual's prescriptions for that drug}
\item{"1e2"}{set to mode for practice's prescriptions for that drug}
\item{"1e3"}{set to mode for population's prescriptions for that drug}
\item{"1f1"}{use value of individual's next prescription}
\item{"1f2"}{use value of practice's next prescription}
\item{"1f3"}{use value of population's next prescription}
\item{"1g1"}{use value of individual's previous prescription}
\item{"1g2"}{use value of practice's previous prescription}
\item{"1g3"}{use value of population's previous prescription}
}}
}
\value{
Dataframe with the same structure as the input
}
\description{
This function accepts a data frame containing prescription information and
replace the implusible quntity by one of the following options.
The user must provide a data input with the following column names:
patid, prodcode, qty (quantity), implausible_qty and pracid (if practice level
information is used for replacement). The implausible_qty column will take value 1
if the given qty in the data is not a plausible value and 0 otherwise. If this column
not provided then only decision 1a (do nothing) can be chosen.
}
