% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec1_implausible_qty}
\alias{dec1_implausible_qty}
\title{implausible  quantity}
\usage{
dec1_implausible_qty(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containg prescription information}
}
\value{
data.frame
}
\description{
This function acepts a data frame containg prodcode, implausible indicator and replace the implusible quntity by one of the following options.
}
