% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec2_missing_qty.R
\name{dec2_missing_qty}
\alias{dec2_missing_qty}
\title{Missing  quantity}
\usage{
dec2_missing_qty(dataset1 = NULL, decision)
}
\arguments{
\item{dataset, }{decsion}
}
\value{
dataset
}
\description{
This function acepts a data frame containg prodcode, and quantity and
replace the missing quntity by one of the following options.
options are  2a  Leave as missing (implicitly drop this prescription)
2b1  set to mean for individual's prescriptions for that drug
2b2  set to mean for practice's prescriptions for that drug
2b3  set to mean for populations's prescriptions for that drug
2c1  set to median for individual's prescriptions for that drug
2c2  set to median for practice's prescriptions for that drug
2c3  set to median for population's prescriptions for that drug
2d1  set to mode for individual's prescriptions for that drug
2d2  set to mode for practice's prescriptions for that drug
2d3  set to mode for population's prescriptions for that drug
2e1  use value of individual's next prescription
2e2  use value of practice's next prescription
2e3  use value of population's next prescription
2f1  use value of individual's previous prescription
2f2  use value of practice's previous prescription
2f3  use value of population's previous prescription
}
