% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec3_implausible_ndd}
\alias{dec3_implausible_ndd}
\title{Implusible ndd}
\usage{
dec3_implausible_ndd(data = NULL, decision)
}
\arguments{
\item{data}{a data frame containing prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"3a"}{use implausible value}
\item{"3b"}{set to missing}
\item{"3c1"}{set to mean for individual's prescriptions for that drug}
\item{"3c2"}{set to mean for practice's prescriptions for that drug}
\item{"3c3"}{set to mean for populations's prescriptions for that drug}
\item{"3d1"}{set to median for individual's prescriptions for that drug}
\item{"3d2"}{set to median for practice's prescriptions for that drug}
\item{"3d3"}{set to median for population's prescriptions for that drug}
\item{"3e1"}{set to mode for individual's prescriptions for that drug}
\item{"3e2"}{set to mode for practice's prescriptions for that drug}
\item{"3e3"}{set to mode for population's prescriptions for that drug}
\item{"3f1"}{use value of individual's next prescription}
\item{"3f2"}{use value of practice's next prescription}
\item{"3f3"}{use value of population's next prescription}
\item{"3g1"}{use value of individual's previous prescription}
\item{"3g2"}{use value of practice's previous prescription}
\item{"3g3"}{use value of population's previous prescription}
}}
}
\value{
Dataframe with the same structure as the input
}
\description{
This function accepts a data frame containing prescription information
and replace the implusible ndd by one of the following options.
The user must provide a data input with the following column names:
patid, prodcode, ndd, implausible_ndd and pracid (if practice level
information is used for replacement). The implausible_ndd column will take value 1
if the given ndd in the data is not a plausible value and 0 otherwise. If this column
not provided then only decision 1a (do nothing) can be chosen.
}
