% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec3_implausible_ndd}
\alias{dec3_implausible_ndd}
\title{Implusible  ndd}
\usage{
dec3_implausible_ndd(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containg prescription information}
}
\value{
data.frame
}
\description{
This function acepts a data frame containg prodcode, ndd, and implusible ndd indicator and replace the implusible ndd by one of the following options.
}
