% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec6_select_stop_date.R
\name{dec6_select_stop_date}
\alias{dec6_select_stop_date}
\title{Stop date
Select which stop date to use
options are	6a	stop1 (start + numdays)
6b	stop2 (start + dose_duration)
6c	stop3 (start + qty/ndd)
If only one stop available, use it
if two available and equal, use that date
if two available and unequal (but within x days), use mean
if three available and unequal, use mean of closest 2 if within x days
6d_15	x = 15
6d_30	x = 30
6d_60	x = 60
6d_90	x = 90
6e 	x = something very big, like 9,999,999}
\usage{
dec6_select_stop_date(dataset1, decision)
}
\arguments{
\item{dataset, }{decsion}
}
\value{
dataset
}
\description{
Stop date
Select which stop date to use
options are	6a	stop1 (start + numdays)
6b	stop2 (start + dose_duration)
6c	stop3 (start + qty/ndd)
If only one stop available, use it
if two available and equal, use that date
if two available and unequal (but within x days), use mean
if three available and unequal, use mean of closest 2 if within x days
6d_15	x = 15
6d_30	x = 30
6d_60	x = 60
6d_90	x = 90
6e 	x = something very big, like 9,999,999
}
