% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec6_select_stop_date}
\alias{dec6_select_stop_date}
\title{Stop date
Select which stop date to use}
\usage{
dec6_select_stop_date(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containg prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"6a"}{stop1 (start + numdays)}
\item{"6b"}{stop2 (start + dose_duration)}
\item{"6c"}{stop3 (start + qty/ndd)}
}}
}
\value{
data.frame
}
\description{
Stop date
Select which stop date to use
}
