% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec7_missing_stop_date.R
\name{dec7_missing_stop_date}
\alias{dec7_missing_stop_date}
\title{Handle missing stop dates
options are	7a	Leave as missing, drop prescription
7b	Use mean for that drug for that individual
7c	Use mean for that drug for all individuals
7d 	Use individual mean for that drug but if not available use population mean}
\usage{
dec7_missing_stop_date(dataset1, decision)
}
\arguments{
\item{dataset, }{decsion}
}
\value{
dataset
}
\description{
Handle missing stop dates
options are	7a	Leave as missing, drop prescription
7b	Use mean for that drug for that individual
7c	Use mean for that drug for all individuals
7d 	Use individual mean for that drug but if not available use population mean
}
