% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec7_missing_stop_date}
\alias{dec7_missing_stop_date}
\title{Missing stop date}
\usage{
dec7_missing_stop_date(data = NULL, decision)
}
\arguments{
\item{data}{a data frame containing prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"7a"}{Leave as missing, drop prescription}
\item{"7b"}{Use mean for that drug for that individual}
\item{"7c"}{Use mean for that drug for all individuals}
\item{"7d"}{Use individual mean for that drug but if not available use population mean}

}}
}
\value{
Dataframe with the same structure as the input
}
\description{
This function identify missing stop dates and perform the following tasks depending on
the given decision. This function can only be called after dec6_select_stop_date is
executed. The input dataset must have the following columns: patid, prodcode, real_stop,
and new_duration.
}
