% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec8_multipleprescription_same_start_date}
\alias{dec8_multipleprescription_same_start_date}
\title{Handle multiple prescriptions for same product on same day}
\usage{
dec8_multipleprescription_same_start_date(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containg prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"8a"}{do nothing}
\item{"8b"}{use mean ndd and mean length}
\item{"8c"}{choose shortest prescription}
\item{"8d"}{choose longest prescription}
\item{"8e"}{sum durations}
}}
}
\value{
Dataframe with the same structure as the input
}
\description{
In situations where rows of identical prescription for the same individual at the same time
is found in the data, this function removes the duplicated prescriptions by choosing one of the
following options. The input data must have the following columns: patid,prodcode,start, real_stop,
and ndd. It's adviced that this function is called after dec6_select_stop_date is executed.
}
