% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec9_overlaping_prescription}
\alias{dec9_overlaping_prescription}
\title{Handle overlapping prescriptions}
\usage{
dec9_overlaping_prescription(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containing prescription information}

\item{decision}{a character specifying the decison to consider for processing
\itemize{
\item{"9a"}{do nothing, allow prescriptions to overlap (implicity sum doses)}
\item{"9b"}{move later prescription to next available time that this product is not prescribed}
}}
}
\value{
Dataframe with the same structure as the input.
}
\description{
In situations where one prescription start before the end date of previous prescription
of the same type (prodcode) for an individual, this function either moves the start date
of the prescription to the end date of the previous prescription or sums the ndd of the two
prescrption at the time overlap depending on the users decision.
The input data must have the following columns: patid,prodcode,start, real_stop,
and ndd.
}
