% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{dec9_overlaping_prescription}
\alias{dec9_overlaping_prescription}
\title{Handle overlapping prescriptions
options are:	9a  do nothing, allow prescriptions to overlap (implicity sum doses)
9b  move later prescription to next available time that this product is not prescribed}
\usage{
dec9_overlaping_prescription(dataset1 = NULL, decision)
}
\arguments{
\item{dataset1}{a data frame containg prescription information}

\item{decision}{a character specifying the decison to consider for processing}
}
\value{
data.frame
}
\description{
Handle overlapping prescriptions
options are:	9a  do nothing, allow prescriptions to overlap (implicity sum doses)
9b  move later prescription to next available time that this product is not prescribed
}
