% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffreduce.R
\name{ffreduce}
\alias{ffreduce}
\title{Utility function to sequentially apply function lists}
\usage{
ffreduce(value, function_list, decision)
}
\arguments{
\item{value}{value}

\item{function_list}{list of decison nodes for processing}

\item{decision}{decision vector}
}
\value{
Dataframe with the same structure as the input
}
\description{
This function accepts the list of functions and decisons corresponding to each
functions and sequentially apply them to the data. That is the resulting data of
function x will be the input for function x+1
}
