% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffreduce.R, R/run.drugPREP.universe.R
\name{ffreduce}
\alias{ffreduce}
\title{freduce
modified version of reduce function in r}
\usage{
ffreduce(value, function_list, decision)

ffreduce(value, function_list, decision)
}
\arguments{
\item{value, }{function_list, decision}

\item{function_list}{list of function}

\item{decision}{decision vector}

\item{value}{value}
}
\value{
value

value
}
\description{
freduce
modified version of reduce function in r

freduce
modified version of reduce function in r
}
