% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implusible_values.R
\name{Implusible_values}
\alias{Implusible_values}
\title{Define implusible}
\usage{
Implusible_values(dataset1 = NULL, min_max_dat = NULL)
}
\arguments{
\item{dataset1}{a data frame object containing the prescription information}

\item{min_max_dat}{a data frame object containing minimum and maximum possible values for each drug}
}
\value{
dataset
}
\description{
Define implusible
}
