% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Implusible_values.R
\name{Implusible_values}
\alias{Implusible_values}
\title{create a shiny that will be populated by the user with min max values
That is data set Immp_val will be generated from dataset1 by removing duplicate productnames
and will have column Prdductcode, productname, qty_min, qty_max, ndd_min, ndd_max
The last 4 columns will be populated by the user
There should be a way to populate Immp_val by the user- rshiny is one way or just using edit(Immp_val)}
\usage{
Implusible_values(dataset1)
}
\value{
dataset
}
\description{
create a shiny that will be populated by the user with min max values
That is data set Immp_val will be generated from dataset1 by removing duplicate productnames
and will have column Prdductcode, productname, qty_min, qty_max, ndd_min, ndd_max
The last 4 columns will be populated by the user
There should be a way to populate Immp_val by the user- rshiny is one way or just using edit(Immp_val)
}
