% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.MVD.R
\name{run.drugPREP.MVD}
\alias{run.drugPREP.MVD}
\title{run drug prep multiverse
outer function to run all decisons
This function should accept multiple decisions for each step and
create multiverse data
some of the computation inside this function will be taken out}
\usage{
run.drugPREP.MVD(dataset1 = NULL, decisions = NULL)
}
\arguments{
\item{dataset, }{decisons,}
}
\value{
dataset
}
\description{
run drug prep multiverse
outer function to run all decisons
This function should accept multiple decisions for each step and
create multiverse data
some of the computation inside this function will be taken out
}
