% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.R
\name{run.drugPREP}
\alias{run.drugPREP}
\title{run drug prep
outer function to run all decisons
some of the computation inside this function will be taken out}
\usage{
run.drugPREP(dataset1 = NULL, decisions = NULL)
}
\arguments{
\item{dataset, }{decisons,}
}
\value{
dataset
}
\description{
run drug prep
outer function to run all decisons
some of the computation inside this function will be taken out
}
