% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.R
\name{run.drugPREP}
\alias{run.drugPREP}
\title{run drug prep
outer function to run all decisons
If you have implusible values to take action, run first the implusible_value function
else you need to choose 'do nothing' for dec1 and dec3}
\usage{
run.drugPREP(dataset1 = NULL, decisions = NULL)
}
\arguments{
\item{dataset, }{decisons,}
}
\value{
dataset
}
\description{
run drug prep
outer function to run all decisons
If you have implusible values to take action, run first the implusible_value function
else you need to choose 'do nothing' for dec1 and dec3
}
