% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{run.drugPREP}
\alias{run.drugPREP}
\title{run drug prep}
\usage{
run.drugPREP(dataset1 = NULL, decisions = NULL)
}
\arguments{
\item{dataset1}{a data frame containing prescription information}

\item{decisions}{a character vector containing list of decison to be taken
\itemize{
\item{"decison1"}{ implussible qantity}
\itemize{
\item{"1a"}{use implausible value}
\item{"1b"}{set to missing}
\item{"1c1"}{set to mean for individual's prescriptions for that drug}
\item{"1c2"}{set to mean for practice's prescriptions for that drug}
\item{"1c3"}{set to mean for populations's prescriptions for that drug}
\item{"1d1"}{set to median for individual's prescriptions for that drug}
\item{"1d2"}{set to median for practice's prescriptions for that drug}
\item{"1d3"}{set to median for population's prescriptions for that drug}
\item{"1e1"}{set to mode for individual's prescriptions for that drug}
\item{"1e2"}{set to mode for practice's prescriptions for that drug}
\item{"1e3"}{set to mode for population's prescriptions for that drug}
\item{"1f1"}{use value of individual's next prescription}
\item{"1f2"}{use value of practice's next prescription}
\item{"1f3"}{use value of population's next prescription}
\item{"1g1"}{use value of individual's previous prescription}
\item{"1g2"}{use value of practice's previous prescription}
\item{"1g3"}{use value of population's previous prescription}
}
\item{"decsion2"}{ Missing quantity}
\itemize{
\item{"2a"}{Leave as missing (implicitly drop this prescription)}
\item{"2b1"}{set to mean for individual's prescriptions for that drug}
\item{"2b2"}{set to mean for practice's prescriptions for that drug}
\item{"2b3"}{set to mean for populations's prescriptions for that drug}
\item{"2c1"}{set to median for individual's prescriptions for that drug}
\item{"2c2"}{set to median for practice's prescriptions for that drug}
\item{"2c3"}{set to median for population's prescriptions for that drug}
\item{"2d1"}{set to mode for individual's prescriptions for that drug}
\item{"2d2"}{set to mode for practice's prescriptions for that drug}
\item{"2d3"}{set to mode for population's prescriptions for that drug}
\item{"2e1"}{use value of individual's next prescription}
\item{"2e2"}{use value of practice's next prescription}
\item{"2e3"}{use value of population's next prescription}
\item{"2f1"}{use value of individual's previous prescription}
\item{"2f2"}{use value of practice}'s previous prescription
\item{"2f3"}{use value of population's previous prescription}
}
\item{"decison3"}{Implausible ndd}
\itemize{
\item{"3a"}{use implausible value}
\item{"3b"}{set to missing}
\item{"3c1"}{set to mean for individual's prescriptions for that drug}
\item{"3c2"}{set to mean for practice's prescriptions for that drug}
\item{"3c3"}{set to mean for populations's prescriptions for that drug}
\item{"3d1"}{set to median for individual's prescriptions for that drug}
\item{"3d2"}{set to median for practice's prescriptions for that drug}
\item{"3d3"}{set to median for population's prescriptions for that drug}
\item{"3e1"}{set to mode for individual's prescriptions for that drug}
\item{"3e2"}{set to mode for practice's prescriptions for that drug}
\item{"3e3"}{set to mode for population's prescriptions for that drug}
\item{"3f1"}{use value of individual's next prescription}
\item{"3f2"}{use value of practice's next prescription}
\item{"3f3"}{use value of population's next prescription}
\item{"3g1"}{use value of individual's previous prescription}
\item{"3g2"}{use value of practice's previous prescription}
\item{"3g3"}{use value of population's previous prescription}
}
\item{"Missing qty}
\itemize{
\item{"4a"}{Leave as missing (implicitly drop this prescription)}
\item{"4b1"}{set to mean for individual's prescriptions for that drug}
\item{"4b2"}{set to mean for practice's prescriptions for that drug}
\item{"4b3"}{set to mean for populations's prescriptions for that drug}
\item{"4c1"}{set to median for individual's prescriptions for that drug}
\item{"4c2"}{set to median for practice's prescriptions for that drug}
\item{"4c3"}{set to median for population's prescriptions for that drug}
\item{"4d1"}{set to mode for individual's prescriptions for that drug}
\item{"4d2"}{set to mode for practice's prescriptions for that drug}
\item{"4d3"}{set to mode for population's prescriptions for that drug}
\item{"4e1"}{use value of individual's next prescription}
\item{"4e2"}{use value of practice's next prescription}
\item{"4e3"}{use value of population's next prescription}
\item{"4f1"}{use value of individual's previous prescription}
\item{"4f2"}{use value of practice}'s previous prescription
\item{"4f3"}{use value of population's previous prescription}
}
}}
}
\value{
data.frame
}
\description{
Utility function to run all decisions sequentially.
If you have implusible values (i.e., implusible quantity and ndd) to take action,
run first the implusible_value function.
If not, you need to choose 'do nothing' for dec1 and dec3.
}
