% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{run.drugPREP}
\alias{run.drugPREP}
\title{run drug prep}
\usage{
run.drugPREP(dataset1 = NULL, decisions = NULL)
}
\arguments{
\item{dataset1}{a data frame containing prescription information}

\item{decisions}{a character vector containing list of decison to be taken}
}
\value{
data.frame
}
\description{
outer function to run all decisons. If you have implusible values to take action, run first the implusible_value function, else you need to choose 'do nothing' for dec1 and dec3
}
