% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.drugPREP.universe.R
\name{shift_interval}
\alias{shift_interval}
\title{shift_interval}
\usage{
shift_interval(x)
}
\arguments{
\item{x}{a data frame containg prescription start and stop dates}
}
\value{
Dataframe with the same structure as the input
}
\description{
shift_interval
}
